Autogen install path based on prefix location
If the environment is autodetected as being from a known
launcher (at the time of coding, Steam is the only one that
is to be supported), generate an install location based on
the detected prefix. This way, if/when installer chain-launching
becomes a thing, the game will be installed in a C:-relative location.

This is actually important because the Genshin, Houkai 3rd
and HSR installers have an issue where they will mis-detect
free space available on the target drive if the game is launched
via "Z:", which is aliased to / and very much a weird thing
for free-space detection. Auto-generating a path into the prefix,
at least for Steam but maybe ultimately coming bac to it for our
own installs as well, works around this WINE issue.